thermal: move `Max31790State` into `control`
Seemed nicer to have it be in the `control` module. I also added some
comments.
