+++ task/thermal/src/main.rs
@@ -38,7 +38,7 @@ use crate::{
     control::ThermalControl,
 };
 use drv_i2c_api::ResponseCode;
-use drv_i2c_devices::max31790::{I2cWatchdog, Max31790};
+use drv_i2c_devices::max31790::I2cWatchdog;
 use idol_runtime::{NotificationHandler, RequestError};
 use ringbuf::*;
 use task_sensor_api::{Sensor as SensorApi, SensorId};
@@ -92,71 +92,6 @@ enum Trace {
 counted_ringbuf!(Trace, 32, Trace::None);
 
 ////////////////////////////////////////////////////////////////////////////////
-
-pub(crate) struct Max31790State {
-    max31790: Max31790,
-    initialized: bool,
-}
-
-impl Max31790State {
-    pub(crate) fn new(max31790: Max31790) -> Self {
-        let mut this = Self {
-            max31790,
-            initialized: false,
-        };
-        // When we first start up, try to initialize the fan controller a few
-        // times, in case there's a transient I2C error.
-        for remaining in (0..3).rev() {
-            if this.initialize().is_ok() {
-                break;
-            }
-            ringbuf_entry!(Trace::FanControllerInitRetry { remaining });
-        }
-        this
-    }
-
-    /// Access the fan controller, attempting to initialize it if it has not yet
-    /// been initialized.
-    #[inline]
-    pub(crate) fn try_initialize(
-        &mut self,
-    ) -> Result<&mut Max31790, ControllerInitError> {
-        if self.initialized {
-            return Ok(&mut self.max31790);
-        }
-
-        self.initialize()
-    }
-
-    // Slow path that actually performs initialization. This is "outlined" so
-    // that we can avoid pushing a stack frame in the case where we just need to
-    // check a bool and return a pointer.
-    #[inline(never)]
-    fn initialize(&mut self) -> Result<&mut Max31790, ControllerInitError> {
-        self.max31790.initialize().map_err(|e| {
-            ringbuf_entry!(Trace::FanControllerInitError(e));
-            ControllerInitError(e)
-        })?;
-
-        self.initialized = true;
-        ringbuf_entry!(Trace::FanControllerInitialized);
-        Ok(&mut self.max31790)
-    }
-}
-
-pub(crate) struct ControllerInitError(ResponseCode);
-
-impl From<ControllerInitError> for ThermalError {
-    fn from(_: ControllerInitError) -> Self {
-        ThermalError::FanControllerUninitialized
-    }
-}
-
-impl From<ControllerInitError> for SensorReadError {
-    fn from(ControllerInitError(code): ControllerInitError) -> Self {
-        SensorReadError::I2cError(code)
-    }
-}
 
 struct ServerImpl<'a> {
     mode: ThermalMode,
